split EquatableMacro into smaller files
