+++ Sources/EquatableMacros/EquatableMacro.swift
@@ -145,7 +145,7 @@ import SwiftSyntaxMacros
 ///  ```
 ///
 public struct EquatableMacro: ExtensionMacro {
-    private static let skippablePropertyWrappers: Set = [
+    static let skippablePropertyWrappers: Set = [
         "AccessibilityFocusState",
         "AppStorage",
         "Bindable",
@@ -248,313 +248,5 @@ public struct EquatableMacro: ExtensionMacro {
         } else {
             return [extensionSyntax]
         }
-    }
-}
-
-extension EquatableMacro {
-    private enum Isolation {
-        case nonisolated
-        case isolated
-        case main
-
-        var keyword: String {
-            switch self {
-            case .nonisolated: "nonisolated"
-            case .isolated: ""
-            case .main: "@MainActor "
-            }
-        }
-    }
-
-    private static func extractIsolation(from node: AttributeSyntax) -> Isolation? {
-        guard let arguments = node.arguments?.as(LabeledExprListSyntax.self) else {
-            return nil
-        }
-
-        for argument in arguments {
-            if argument.label?.text == "isolation" {
-                if let memberAccess = argument.expression.as(MemberAccessExprSyntax.self) {
-                    switch memberAccess.declName.baseName.text {
-                    case "isolated": return .isolated
-                    case "nonisolated": return .nonisolated
-                    #if swift(>=6.2)
-                        case "main": return .main
-                    #endif
-                    default: return nil
-                    }
-                }
-            }
-        }
-
-        return nil
-    }
-
-    // Skip properties with SwiftUI attributes (like @State, @Binding, etc.) or if they are marked with @EqutableIgnored
-    private static func shouldSkip(_ varDecl: VariableDeclSyntax) -> Bool {
-        varDecl.attributes.contains { attribute in
-            if let atribute = attribute.as(AttributeSyntax.self),
-               shouldSkip(atribute: atribute) {
-                return true
-            }
-            return false
-        }
-    }
-
-    private static func shouldSkip(atribute node: AttributeSyntax) -> Bool {
-        if let identifierType = node.attributeName.as(IdentifierTypeSyntax.self),
-           shouldSkip(identifierType: identifierType) {
-            return true
-        }
-        if let memberType = node.attributeName.as(MemberTypeSyntax.self),
-           Self.shouldSkip(memberType: memberType) {
-            return true
-        }
-        return false
-    }
-
-    private static func shouldSkip(identifierType node: IdentifierTypeSyntax) -> Bool {
-        if node.name.text == "EquatableIgnored" {
-            return true
-        }
-        if skippablePropertyWrappers.contains(node.name.text) {
-            return true
-        }
-        return false
-    }
-
-    private static func shouldSkip(memberType node: MemberTypeSyntax) -> Bool {
-        if node.baseType.as(IdentifierTypeSyntax.self)?.name.text == "SwiftUI",
-           skippablePropertyWrappers.contains(node.name.text) {
-            return true
-        }
-        return false
-    }
-
-    private static func isMarkedWithEquatableIgnoredUnsafeClosure(_ varDecl: VariableDeclSyntax) -> Bool {
-        varDecl.attributes.contains(where: { attribute in
-            if let attributeName = attribute.as(AttributeSyntax.self)?.attributeName.as(IdentifierTypeSyntax.self)?.name.text {
-                return attributeName == "EquatableIgnoredUnsafeClosure"
-            }
-            return false
-        })
-    }
-
-    private static func compare(lhs: (name: String, type: TypeSyntax?), rhs: (name: String, type: TypeSyntax?)) -> Bool {
-        // "id" always comes first
-        if lhs.name == "id" { return true }
-        if rhs.name == "id" { return false }
-
-        let lhsComplexity = typeComplexity(lhs.type)
-        let rhsComplexity = typeComplexity(rhs.type)
-
-        if lhsComplexity == rhsComplexity {
-            return lhs.name < rhs.name
-        }
-        return lhsComplexity < rhsComplexity
-    }
-
-    // swiftlint:disable:next cyclomatic_complexity
-    private static func typeComplexity(_ type: TypeSyntax?) -> Int {
-        guard let type else { return 100 } // Unknown types go last
-
-        let typeString = type.description.trimmingCharacters(in: CharacterSet.whitespacesAndNewlines)
-
-        switch typeString {
-        case "Bool": return 1
-        case "Int", "Int8", "Int16", "Int32", "Int64": return 2
-        case "UInt", "UInt8", "UInt16", "UInt32", "UInt64": return 3
-        case "Float", "Double": return 4
-        case "String": return 5
-        case "Character": return 6
-        case "Date": return 7
-        case "Data": return 8
-        case "URL": return 9
-        case "UUID": return 10
-        default:
-            if type.is(OptionalTypeSyntax.self) {
-                if let wrappedType = type.as(OptionalTypeSyntax.self)?.wrappedType {
-                    return typeComplexity(wrappedType) + 20
-                }
-            }
-
-            if type.isArray {
-                return 30
-            }
-
-            if type.isDictionary {
-                return 40
-            }
-
-            return 50
-        }
-    }
-
-    private static func makeClosureDiagnostic(for varDecl: VariableDeclSyntax) -> Diagnostic {
-        let attribute = AttributeSyntax(
-            leadingTrivia: .space,
-            atSign: .atSignToken(),
-            attributeName: IdentifierTypeSyntax(name: .identifier("EquatableIgnoredUnsafeClosure")),
-            trailingTrivia: .space
-        )
-        let existingAttributes = varDecl.attributes
-        let newAttributes = existingAttributes + [.attribute(attribute.with(\.leadingTrivia, .space))]
-        let fixedDecl = varDecl.with(\.attributes, newAttributes)
-        let diagnostic = Diagnostic(
-            node: varDecl,
-            message: MacroExpansionErrorMessage("Arbitary closures are not supported in @Equatable"),
-            fixIt: .replace(
-                message: SimpleFixItMessage(
-                    message: """
-                    Consider marking the closure with\
-                    @EquatableIgnoredUnsafeClosure if it doesn't effect the view's body output.
-                    """,
-                    fixItID: MessageID(
-                        domain: "",
-                        id: "test"
-                    )
-                ),
-                oldNode: varDecl,
-                newNode: fixedDecl
-            )
-        )
-
-        return diagnostic
-    }
-
-    private static func generateEquatableExtensionSyntax(
-        sortedProperties: [(name: String, type: TypeSyntax?)],
-        type: TypeSyntaxProtocol,
-        isolation: Isolation
-    ) -> ExtensionDeclSyntax? {
-        guard !sortedProperties.isEmpty else {
-            let extensionDecl: DeclSyntax = switch isolation {
-            case .nonisolated:
-                """
-                extension \(type): Equatable {
-                    nonisolated public static func == (lhs: \(type), rhs: \(type)) -> Bool {
-                        true
-                    }
-                }
-                """
-            case .isolated:
-                """
-                extension \(type): Equatable {
-                    public static func == (lhs: \(type), rhs: \(type)) -> Bool {
-                        true
-                    }
-                }
-                """
-            case .main:
-                """
-                extension \(type): @MainActor Equatable {
-                    public static func == (lhs: \(type), rhs: \(type)) -> Bool {
-                        true
-                    }
-                }
-                """
-            }
-
-            return extensionDecl.as(ExtensionDeclSyntax.self)
-        }
-
-        let comparisons = sortedProperties.map { property in
-            "lhs.\(property.name) == rhs.\(property.name)"
-        }.joined(separator: " && ")
-
-        let equalityImplementation = comparisons.isEmpty ? "true" : comparisons
-
-        let extensionDecl: DeclSyntax = switch isolation {
-        case .nonisolated:
-            """
-            extension \(type): Equatable {
-                nonisolated public static func == (lhs: \(type), rhs: \(type)) -> Bool {
-                    \(raw: equalityImplementation)
-                }
-            }
-            """
-        case .isolated:
-            """
-            extension \(type): Equatable {
-                public static func == (lhs: \(type), rhs: \(type)) -> Bool {
-                    \(raw: equalityImplementation)
-                }
-            }
-            """
-        case .main:
-            """
-            extension \(type): @MainActor Equatable {
-                public static func == (lhs: \(type), rhs: \(type)) -> Bool {
-                    \(raw: equalityImplementation)
-                }
-            }
-            """
-        }
-
-        return extensionDecl.as(ExtensionDeclSyntax.self)
-    }
-
-    private static func generateHashableExtensionSyntax(
-        sortedProperties: [(name: String, type: TypeSyntax?)],
-        type: TypeSyntaxProtocol,
-        isolation: Isolation
-    ) -> ExtensionDeclSyntax? {
-        guard !sortedProperties.isEmpty else {
-            let hashableExtensionDecl: DeclSyntax = switch isolation {
-            case .nonisolated:
-                """
-                extension \(raw: type) {
-                    nonisolated public func hash(into hasher: inout Hasher) {}
-                }
-                """
-            case .isolated:
-                """
-                extension \(raw: type) {
-                    public func hash(into hasher: inout Hasher) {}
-                }
-                """
-            case .main:
-                """
-                extension \(raw: type) {
-                    public func hash(into hasher: inout Hasher) {}
-                }
-                """
-            }
-
-            return hashableExtensionDecl.as(ExtensionDeclSyntax.self)
-        }
-
-        let hashableImplementation = sortedProperties.map { property in
-            "hasher.combine(\(property.name))"
-        }
-        .joined(separator: "\n")
-
-        let hashableExtensionDecl: DeclSyntax = switch isolation {
-        case .nonisolated:
-            """
-            extension \(raw: type) {
-                nonisolated public func hash(into hasher: inout Hasher) {
-                    \(raw: hashableImplementation)
-                }
-            }
-            """
-        case .isolated:
-            """
-            extension \(raw: type) {
-                public func hash(into hasher: inout Hasher) {
-                    \(raw: hashableImplementation)
-                }
-            }
-            """
-        case .main:
-            """
-            extension \(raw: type) {
-                public func hash(into hasher: inout Hasher) {
-                    \(raw: hashableImplementation)
-                }
-            }
-            """
-        }
-
-        return hashableExtensionDecl.as(ExtensionDeclSyntax.self)
     }
 }
